ELEC-198: Renamed FSM files
