+++ project/driver_controls/src/direction_fsm.c
@@ -1,4 +1,4 @@
-#include "direction_state.h"
+#include "direction_fsm.h"
 #include "input_event.h"
 
 // Direction selector FSM state definitions
